Move filter_structures_not_present_in_annotation to structures.py
